pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "switch-inventory-tool"
version = "0.1.0"
description = "Network Switch Inventory Management Tool"
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "openpyxl>=3.0.7", 
    "PyYAML>=5.4.0",
    "pytest>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_paths = ["src", "."]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["src"]
